refactor(node/state_delta): extract buffering submodule (increment 3)
Continues the state_delta god-file split. Moves the governance-pending
drain / absorb / recover / fence cluster (~765 lines) into `buffering.rs`
— behavior-preserving, pure code movement (mod.rs: +10 wiring lines, -766).

Moved:
- drain_governance_pending, drain_all_governance_pending
- drain_absorbed, drain_absorbed_leaves, drain_absorbed_records,
  drain_all_absorbed
- recover_absorbed_records, recover_absorbed_on_startup
- FenceOutcome + fence_and_maybe_absorb

Wiring: external entry points (drain_all_governance_pending,
drain_all_absorbed, recover_absorbed_on_startup — called from sync::manager
and network_event::namespace) are re-exported `pub(crate)` to preserve the
`crate::handlers::state_delta::*` paths; the apply-path/test helpers are
re-imported so call sites and the in-module test suite are unchanged.
Explicit imports (no `use super::*`), per the established convention.

Validated: node lib tests (304) and sync_sim (314) pass.

Remaining increment: delta-store setup (init_delta_store /
choose_owned_identity / DeltaStoreSetup). The two orchestrators
(handle_state_delta, apply_authorized_state_delta) stay in mod.rs.
